feat(incomplet): translation transformation and polarization

+++ image_interpolation.py
@@ -3,7 +3,7 @@ import numpy as np
 def interpolate_by_nearest_neighbor(transformed_pixels, img_size, h_scale, w_scale):
 
     height, width = img_size
-    new_img_matrix = np.full((height * h_scale, width * w_scale), 0, dtype=np.uint8)
+    new_img_matrix = np.full((int(height * h_scale), int(width * w_scale)), 0, dtype=np.uint8)
 
     for pixel in transformed_pixels:
         v, w, value = pixel
